# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing community waste collection — households, pickups, payments, reports.",
        "title": "Community Waste Collection API",
        "contact": {
            "name": "Muhammad Faisal Affan"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/payments": {
            "get": {
                "description": "List payments with filters (status, household, date range)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "List payments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by household",
                        "name": "household_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "From date (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To date (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a manual payment linked to a household",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create payment",
                "parameters": [
                    {
                        "description": "Payment data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.Payment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/payments/{id}/confirm": {
            "put": {
                "description": "Confirm payment with proof file upload (BR-06)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Confirm payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Payment proof file",
                        "name": "proof_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.Payment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_faisalaffan_community-waste-collection-api_internal_domain.CreatePaymentRequest": {
            "type": "object",
            "required": [
                "household_id"
            ],
            "properties": {
                "household_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faisalaffan_community-waste-collection-api_internal_domain.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "household_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_date": {
                    "type": "string"
                },
                "proof_file_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "waste_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.apiError"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.pagination"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.ValidationDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.ValidationDetail"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    }
}